Fix: Read and Secure Read should check byte count prior to raising message received event. This would otherwise lead to false events being raised on disconnect.
Fix: Remembering client and host info should occur prior to reading incoming data. This would otherwise lead to KeyNotFound exceptions when a client disconnects on read but before the client was registered.

Updated Dispose loop to use less than operator and removed the -1 after the count to simply the operation.

Updated RemoveClient function to include more lines in the try / catch block.

Removed Log line on ObjectDisposedException for the AcceptTcpClient. This is an excpected condition when the TCP Listener is stopped.